Cleanup files for the MPD geometry, added new config files for single MPD and ArgonCube + MPD
