+++ duneggd/Active/NDHPgTPC.py
@@ -1,10 +1,10 @@
 #!/usr/bin/env python
 '''
-Builds compontents for the MPT ECAL
+Builds compontents for the MPD
 '''
 
 import gegede.builder
-from duneggd.SubDetector import NDHPgTPC_v02 as NDHPgTPC
+from duneggd.SubDetector import NDHPgTPC as NDHPgTPC
 from gegede import Quantity as Q
 from math import floor, atan, sin, cos, tan, sqrt, pi, asin
 
@@ -217,7 +217,7 @@ class NDHPgTPCDetElementBuilder(gegede.builder.Builder):
                     phi_range = [Q("0deg"), Q("360deg")],
                     material = 'Air',
                     nsides = 8,
-                    nModules = 3,
+                    nModules = 5,
                     output_name = 'MPTECalDetElement',
                     HGlayerType1_builder_name = "NDHPgTPCHGLayerBuilder1",
                     HGlayerType2_builder_name = "NDHPgTPCHGLayerBuilder2",
@@ -228,10 +228,14 @@ class NDHPgTPCDetElementBuilder(gegede.builder.Builder):
                     yokeMaterial = "Steel",
                     yokeThickness = Q("500mm"),
                     yokePhiCutout = Q("90deg"),
-                    rInnerEcal = Q("2740mm"),
-                    Barrel_halfZ = Q('2600mm'),
+                    rInnerTPC = Q("2740mm"),
+                    TPC_halfZ = Q('2600mm'),
                     nLayers = [30, 20],
-                    typeLayers = ['HG', 'LG']
+                    typeLayers = ['HG1', 'LG1'],
+                    magnetMaterial = "Aluminum",
+                    magnetThickness = Q("130mm"),
+                    magnetInnerR = Q("3250cm"),
+                    magnetHalfLength = Q("5m")
                     )
 
     #def configure(self, **kwds):
@@ -264,8 +268,7 @@ class NDHPgTPCDetElementBuilder(gegede.builder.Builder):
         LG_layer_type1_thickness = LGLayerType1_shape.dz * 2
         LG_layer_type2_thickness = LGLayerType2_shape.dz * 2
 
-        safety = Q("1mm")
-        ecal_module_thickness = safety
+        ecal_module_thickness = Q("0mm")
 
         for nlayer, type in zip(self.nLayers, self.typeLayers):
             if type == "HG1":
@@ -281,6 +284,19 @@ class NDHPgTPCDetElementBuilder(gegede.builder.Builder):
 
         return ecal_module_thickness
 
+    def get_pv_endcap_length(self, geom):
+
+        safety = Q("0.1mm")
+        pv_rInner = self.rInnerTPC
+        pvHalfLength = self.TPC_halfZ
+        pv_rmin = sqrt((pv_rInner/Q("1mm"))**2)*Q("1mm")
+        pv_rmax = pv_rmin + self.pvThickness
+
+        rmin_sph = sqrt((pvHalfLength/Q("1mm"))**2 + (pv_rmax/Q("1mm"))**2)*Q("1mm")
+        rmax_sph = rmin_sph + self.pvThickness
+
+        return rmax_sph
+
     def construct(self, geom):
 
         if self.checkVariables() == False:
@@ -293,14 +309,58 @@ class NDHPgTPCDetElementBuilder(gegede.builder.Builder):
             self.construct_ecal_endcap_staves(geom)
         elif self.geometry == 'PV':
             self.construct_pv(geom)
-        elif self.geometry == 'YokeBarrel':
-            self.construct_yoke_barrel_staves(geom)
-        elif self.geometry == 'YokeEndcap':
-            self.construct_yoke_endcap_staves(geom)
+        elif self.geometry == 'Magnet':
+            self.construct_magnet(geom)
+        # elif self.geometry == 'YokeBarrel':
+        #     self.construct_yoke_barrel_staves(geom)
+        # elif self.geometry == 'YokeEndcap':
+        #     self.construct_yoke_endcap_staves(geom)
         else:
             print "Could not find the geometry asked!"
             return
         return
+
+    def construct_magnet(self, geom):
+        ''' construct the Magnet '''
+
+        print "Construct Magnet"
+
+        magnet_name = self.output_name
+        magnet_shape = geom.shapes.Tubs(magnet_name, rmin=self.magnetInnerR, rmax=self.magnetInnerR+self.magnetThickness, dz=self.magnetHalfLength, sphi="0deg", dphi="360deg")
+        magnet_vol = geom.structure.Volume(magnet_name+"_vol", shape=magnet_shape, material=self.magnetMaterial)
+
+        self.add_volume(magnet_vol)
+
+    def construct_pv(self, geom):
+        ''' construct the Pressure Vessel '''
+
+        print "Construct PV Barrel"
+
+        nsides = self.nsides
+        pv_rInner = self.rInnerTPC
+        pvHalfLength = self.TPC_halfZ
+        pv_rmin = sqrt((pv_rInner/Q("1mm"))**2)*Q("1mm")
+        pv_rmax = pv_rmin + self.pvThickness
+
+        # build the pressure vessel barrel
+        pvb_name = self.output_name + "Barrel"
+        pvb_shape = geom.shapes.Tubs(pvb_name, rmin=pv_rmin, rmax=pv_rmax, dz=pvHalfLength, sphi="0deg", dphi="360deg")
+        pvb_vol = geom.structure.Volume(pvb_name+"_vol", shape=pvb_shape, material=self.pvMaterial)
+
+        self.add_volume(pvb_vol)
+
+        print "Construct PV Endcap"
+
+        # build the pressure vessel endcaps
+        pv_endcap_shape_tub = geom.shapes.Tubs("PVEndcap_tubs", rmin=Q("0mm"), rmax=pv_rmax, dz=2*pvHalfLength, sphi="0deg", dphi="360deg")
+        rmin_sph = sqrt((pvHalfLength/Q("1mm"))**2 + (pv_rmax/Q("1mm"))**2)*Q("1mm")
+        rmax_sph = rmin_sph + self.pvThickness
+        pv_endcap_shape_sphere = geom.shapes.Sphere("PVEndcap_sphere", rmin=rmin_sph, rmax=rmax_sph, sphi="0deg", dphi="360deg", stheta="0deg", dtheta="179.99999deg")
+
+        pv_endcap_shape_inter = geom.shapes.Boolean( self.output_name + "Endcap", type='intersection', first=pv_endcap_shape_tub, second=pv_endcap_shape_sphere)
+        pvec_vol = geom.structure.Volume( self.output_name+"Endcap"+"_vol", shape=pv_endcap_shape_inter, material=self.pvMaterial )
+
+        self.add_volume(pvec_vol)
 
     def construct_ecal_barrel_staves(self, geom):
         ''' construct a set of ECAL staves for the Barrel '''
@@ -309,102 +369,55 @@ class NDHPgTPCDetElementBuilder(gegede.builder.Builder):
         #   /----------------\         //Small side
         #  /                  \
         # /____________________\       //Large side
-
+        #
+        # z
+        # ^
+        # |
+        # |-----> x
+        #
         # need to create the layer based on the position in depth z -> different layer sizes
 
         print "Construct ECAL Barrel"
 
         # ECAL Barrel
+        safety = Q("0.1mm")
         nsides = self.nsides
-        ecal_module_thickness = self.get_ecal_module_thickness(geom)
-
-        max_dim_x = 2 * tan( pi/nsides ) * self.rInnerEcal + ecal_module_thickness / sin( 2*pi/nsides ) #large side
-        min_dim_x = max_dim_x - 2*ecal_module_thickness / tan( 2*pi/nsides ) # small side
-
-        Ecal_Barrel_halfZ = self.Barrel_halfZ
-        Ecal_Barrel_n_modules = self.nModules
-        Ecal_Barrel_module_dim = Ecal_Barrel_halfZ * 2 / Ecal_Barrel_n_modules
-
-        X = self.rInnerEcal + ecal_module_thickness / 2.
-        #end of slabs aligned to inner face of support plate in next stave (not the outer surface)
-        Y = (ecal_module_thickness / 2.) / sin(2.*pi/nsides)
         dphi = (2*pi/nsides)
         hphi = dphi/2;
 
-        #Mother volume Barrel
-        barrel_shape = geom.shapes.PolyhedraRegular(self.output_name, numsides=nsides, rmin=self.rInnerEcal, rmax=self.rInnerEcal + ecal_module_thickness, dz=Ecal_Barrel_halfZ)
-        barrel_lv = geom.structure.Volume(self.output_name+"_vol", shape=barrel_shape, material=self.material)
+        #ecal module thickness
+        ecal_module_thickness = self.get_ecal_module_thickness(geom)
+        ecal_module_thickness_noSupport = ecal_module_thickness - safety
+        #inner radius ecal (TPC + pv + safety)
+        rInnerEcal = self.rInnerTPC + self.pvThickness
+        #barrel length (TPC + PV)
+        Barrel_halfZ = self.get_pv_endcap_length(geom)
+        #outer radius ecal (inner radius ecal + ecal module)
+        rOuterEcal = rInnerEcal + ecal_module_thickness
 
-        # #Create a template module
-        # stave_name = self.output_name + "_stave"
-        # stave_volname = self.output_name + "_stave" + "_vol"
-        #
-        # stave_shape = geom.shapes.Trapezoid(stave_name, dx1=max_dim_x/2.0, dx2=min_dim_x/2.0,
-        # dy1=Ecal_Barrel_module_dim/2.0, dy2=Ecal_Barrel_module_dim/2.0,
-        # dz=ecal_module_thickness/2.0)
-        #
-        # stave_lv = geom.structure.Volume(stave_volname, shape=stave_shape, material=self.material)
-        #
-        # zPos = Q("0mm")
-        # layer_id = 1
-        # sensname = self.output_name + "_vol"
-        #
-        # for nlayer, type in zip(self.nLayers, self.typeLayers):
-        #     for ilayer in range(nlayer):
-        #
-        #         if type == 'HG':
-        #             #Configure the layer length based on the zPos in the stave
-        #             HGLayer_builder = self.get_builder(self.HGlayer_builder_name)
-        #             layer_thickness = NDHPgTPCHGLayerBuilder.depth(HGLayer_builder)
-        #             l_dim_x = max_dim_x - 2 * (zPos + layer_thickness) / tan(2.* pi / nsides)
-        #             l_dim_y = Ecal_Barrel_module_dim
-        #             layername = self.output_name + "_stave" + "_module" + "_layer_%02i" % (layer_id)
-        #
-        #             NDHPgTPCHGLayerBuilder.BarrelConfigurationLayer(HGLayer_builder, l_dim_x, l_dim_y, layername, sensname, "Box")
-        #             NDHPgTPCHGLayerBuilder.construct(HGLayer_builder, geom)
-        #             layer_lv = HGLayer_builder.get_volume(layername+"_vol")
-        #
-        #         if type == 'LG':
-        #             LGLayer_builder = self.get_builder(self.LGlayer_builder_name)
-        #             layer_thickness = NDHPgTPCLGLayerBuilder.depth(LGLayer_builder)
-        #             l_dim_x = max_dim_x - 2 * (zPos + layer_thickness) / tan(2.* pi / nsides)
-        #             l_dim_y = Ecal_Barrel_module_dim
-        #             layername = self.output_name + "_stave" + "_module" + "_layer_%02i" % (layer_id)
-        #
-        #             NDHPgTPCLGLayerBuilder.BarrelConfigurationLayer(LGLayer_builder, l_dim_x, l_dim_y, layername, sensname, "Box")
-        #             NDHPgTPCLGLayerBuilder.construct(LGLayer_builder, geom)
-        #             layer_lv = LGLayer_builder.get_volume(layername+"_vol")
-        #
-        #         #Placement layer in stave
-        #         layer_pos = geom.structure.Position(layername+"_pos", z=zPos + layer_thickness/2.0 - ecal_module_thickness/2.0)
-        #         layer_pla = geom.structure.Placement(layername+"_pla", volume=layer_lv, pos=layer_pos)
-        #
-        #         stave_lv.placements.append(layer_pla.name)
-        #
-        #         zPos += layer_thickness;
-        #         layer_id += 1
-        #
-        # #Placing the staves and modules
-        # for istave in range(nsides):
-        #     stave_id = istave+1
-        #     dstave = int( nsides/4.0 )
-        #     phirot =  hphi + pi/2.0
-        #     phirot += (istave - dstave)*dphi
-        #     phirot2 =  (istave - dstave) * dphi + hphi
-        #
-        #     for imodule in range(Ecal_Barrel_n_modules):
-        #         module_id = imodule+1
-        #         print "Placing stave ", stave_id, " and module ", module_id
-        #
-        #         #Placement staves in Barrel
-        #         name = stave_lv.name + "_%i" % (stave_id) + "_module%i" % (module_id)
-        #         pos = geom.structure.Position(name + "_pos", x=(X*cos(phirot2)-Y*sin(phirot2)), y=(( X*sin(phirot2)+Y*cos(phirot2) )), z=( imodule+0.5 )*Ecal_Barrel_module_dim - Ecal_Barrel_halfZ )
-        #         rot = geom.structure.Rotation(name + "_rot", x=pi/2.0, y=phirot+pi, z=Q("0deg"))
-        #         pla = geom.structure.Placement(name + "_pla", volume=stave_lv, pos=pos, rot=rot)
-        #
-        #         barrel_lv.placements.append(pla.name)
-        #
-        # self.add_volume(barrel_lv)
+        #maximum dimension of the stave
+        max_dim_stave = 2 * tan( pi/nsides ) * rInnerEcal + ecal_module_thickness_noSupport / sin( 2*pi/nsides )
+        #minimum dimension of the stave
+        min_dim_stave = max_dim_stave - 2*ecal_module_thickness / tan( 2*pi/nsides )
+
+        Ecal_Barrel_halfZ = Barrel_halfZ
+        Ecal_Barrel_n_modules = self.nModules
+        #dimension of a module along the ND x direction
+        Ecal_Barrel_module_dim = Ecal_Barrel_halfZ * 2 / Ecal_Barrel_n_modules
+
+        print "Large side of the stave", max_dim_stave
+        print "Small side of the stave", min_dim_stave
+        print "Barrel module dim in z", Ecal_Barrel_module_dim
+        print "Module thickness", ecal_module_thickness
+
+        #Position of the stave in the Barrel (local coordinates)
+        X = rInnerEcal + safety + ecal_module_thickness / 2.
+        Y = (ecal_module_thickness_noSupport / 2.) / sin(2.*pi/nsides)
+
+
+        #Mother volume Barrel
+        barrel_shape = geom.shapes.PolyhedraRegular(self.output_name, numsides=nsides, rmin=rInnerEcal, rmax=rOuterEcal+2*safety, dz=Ecal_Barrel_halfZ+safety)
+        barrel_lv = geom.structure.Volume(self.output_name+"_vol", shape=barrel_shape, material=self.material)
 
         sensname = self.output_name + "_vol"
         for istave in range(nsides):
@@ -421,8 +434,8 @@ class NDHPgTPCDetElementBuilder(gegede.builder.Builder):
                 stave_name = self.output_name + "_stave%02i" % (stave_id) + "_module%02i" % (module_id)
                 stave_volname = self.output_name + "_stave%02i" % (stave_id) + "_module%02i" % (module_id) + "_vol"
 
-                stave_shape = geom.shapes.Trapezoid(stave_name, dx1=max_dim_x/2.0, dx2=min_dim_x/2.0,
-                dy1=Ecal_Barrel_module_dim/2.0, dy2=Ecal_Barrel_module_dim/2.0,
+                stave_shape = geom.shapes.Trapezoid(stave_name, dx1=max_dim_stave/2.0, dx2=min_dim_stave/2.0,
+                dy1=(Ecal_Barrel_module_dim-safety)/2.0, dy2=(Ecal_Barrel_module_dim-safety)/2.0,
                 dz=ecal_module_thickness/2.0)
 
                 stave_lv = geom.structure.Volume(stave_volname, shape=stave_shape, material=self.material)
@@ -439,8 +452,8 @@ class NDHPgTPCDetElementBuilder(gegede.builder.Builder):
                             #Configure the layer length based on the zPos in the stave
                             HGLayer_builder = self.get_builder(self.HGlayerType1_builder_name)
                             layer_thickness = NDHPgTPCHGLayerBuilder.depth(HGLayer_builder)
-                            l_dim_x = max_dim_x - 2 * (zPos + layer_thickness) / tan(2.* pi / nsides)
-                            l_dim_y = Ecal_Barrel_module_dim
+                            l_dim_x = max_dim_stave - 2 * (zPos + layer_thickness) / tan(2.* pi / nsides)
+                            l_dim_y = Ecal_Barrel_module_dim - safety
 
                             NDHPgTPCHGLayerBuilder.BarrelConfigurationLayer(HGLayer_builder, l_dim_x, l_dim_y, layername, sensname, "Box")
                             NDHPgTPCHGLayerBuilder.construct(HGLayer_builder, geom)
@@ -450,19 +463,18 @@ class NDHPgTPCDetElementBuilder(gegede.builder.Builder):
                             #Configure the layer length based on the zPos in the stave
                             HGLayer_builder = self.get_builder(self.HGlayerType2_builder_name)
                             layer_thickness = NDHPgTPCHGLayerBuilder.depth(HGLayer_builder)
-                            l_dim_x = max_dim_x - 2 * (zPos + layer_thickness) / tan(2.* pi / nsides)
-                            l_dim_y = Ecal_Barrel_module_dim
+                            l_dim_x = max_dim_stave - 2 * (zPos + layer_thickness) / tan(2.* pi / nsides)
+                            l_dim_y = Ecal_Barrel_module_dim - safety
 
                             NDHPgTPCHGLayerBuilder.BarrelConfigurationLayer(HGLayer_builder, l_dim_x, l_dim_y, layername, sensname, "Box")
                             NDHPgTPCHGLayerBuilder.construct(HGLayer_builder, geom)
                             layer_lv = HGLayer_builder.get_volume(layername+"_vol")
 
-
                         if type == 'LG1':
                             LGLayer_builder = self.get_builder(self.LGlayerType1_builder_name)
                             layer_thickness = NDHPgTPCLGLayerBuilder.depth(LGLayer_builder)
-                            l_dim_x = max_dim_x - 2 * (zPos + layer_thickness) / tan(2.* pi / nsides)
-                            l_dim_y = Ecal_Barrel_module_dim
+                            l_dim_x = max_dim_stave - 2 * (zPos + layer_thickness) / tan(2.* pi / nsides)
+                            l_dim_y = Ecal_Barrel_module_dim - safety
 
                             NDHPgTPCLGLayerBuilder.BarrelConfigurationLayer(LGLayer_builder, l_dim_x, l_dim_y, layername, sensname, "Box")
                             NDHPgTPCLGLayerBuilder.construct(LGLayer_builder, geom)
@@ -471,8 +483,8 @@ class NDHPgTPCDetElementBuilder(gegede.builder.Builder):
                         if type == 'LG2':
                             LGLayer_builder = self.get_builder(self.LGlayerType2_builder_name)
                             layer_thickness = NDHPgTPCLGLayerBuilder.depth(LGLayer_builder)
-                            l_dim_x = max_dim_x - 2 * (zPos + layer_thickness) / tan(2.* pi / nsides)
-                            l_dim_y = Ecal_Barrel_module_dim
+                            l_dim_x = max_dim_stave - 2 * (zPos + layer_thickness) / tan(2.* pi / nsides)
+                            l_dim_y = Ecal_Barrel_module_dim - safety
 
                             NDHPgTPCLGLayerBuilder.BarrelConfigurationLayer(LGLayer_builder, l_dim_x, l_dim_y, layername, sensname, "Box")
                             NDHPgTPCLGLayerBuilder.construct(LGLayer_builder, geom)
@@ -489,7 +501,7 @@ class NDHPgTPCDetElementBuilder(gegede.builder.Builder):
 
                 #Placement staves in Barrel
                 name = stave_lv.name
-                pos = geom.structure.Position(name + "_pos", x=(X*cos(phirot2)-Y*sin(phirot2)), y=(( X*sin(phirot2)+Y*cos(phirot2) )), z=( imodule+0.5 )*Ecal_Barrel_module_dim - Ecal_Barrel_halfZ )
+                pos = geom.structure.Position(name + "_pos", x=(X*cos(phirot2)-Y*sin(phirot2)), y=(( X*sin(phirot2)+Y*cos(phirot2) )), z=( imodule+0.5 )*Ecal_Barrel_module_dim - Barrel_halfZ )
                 rot = geom.structure.Rotation(name + "_rot", x=pi/2.0, y=phirot+pi, z=Q("0deg"))
                 pla = geom.structure.Placement(name + "_pla", volume=stave_lv, pos=pos, rot=rot)
 
@@ -503,12 +515,15 @@ class NDHPgTPCDetElementBuilder(gegede.builder.Builder):
         print "Construct ECAL Endcap"
 
         # ECAL Endcap
+        safety = Q("0.1mm")
         nsides = self.nsides
         ecal_module_thickness = self.get_ecal_module_thickness(geom)
+        rInnerEcal = self.rInnerTPC + self.pvThickness + safety
+        Barrel_halfZ = self.get_pv_endcap_length(geom) + safety
 
         EcalEndcap_inner_radius = Q("0mm")
-        EcalEndcap_outer_radius = self.rInnerEcal + ecal_module_thickness
-        Ecal_Barrel_halfZ = self.Barrel_halfZ
+        EcalEndcap_outer_radius = rInnerEcal + ecal_module_thickness
+        Ecal_Barrel_halfZ = Barrel_halfZ
         EcalEndcap_min_z = Ecal_Barrel_halfZ
         EcalEndcap_max_z = Ecal_Barrel_halfZ + ecal_module_thickness
         Ecal_Barrel_n_modules = self.nModules
@@ -522,86 +537,6 @@ class NDHPgTPCDetElementBuilder(gegede.builder.Builder):
 
         endcap_shape = geom.shapes.Boolean( self.output_name, type='subtraction', first=endcap_shape_max, second=endcap_shape_min )
         endcap_lv = geom.structure.Volume( self.output_name+"_vol", shape=endcap_shape, material=self.material )
-
-        # #Create a template module
-        # endcap_stave_full = geom.shapes.PolyhedraRegular("ECAL_stave_full", numsides=nsides, sphi=pi/8, dphi=Q("360deg"), rmin=rmin, rmax=rmax, dz=ecal_module_thickness)
-        # quadr = rmax
-        # quadrant = geom.shapes.Box("Quadrant", dx=quadr, dy=quadr, dz=ecal_module_thickness/2)
-        #
-        # endcap_stave_pos = geom.structure.Position("EcalEndcapStave"+"_pos", x=quadr, y=quadr, z=Q("0mm"))
-        # endcap_stave_shape = geom.shapes.Boolean("EcalEndcapStave", type='intersection', first=endcap_stave_full, second=quadrant, pos=endcap_stave_pos)
-        # endcap_stave_lv = geom.structure.Volume("EcalEndcapStave"+"_vol", shape=endcap_stave_shape, material=self.material)
-        #
-        # zPos = Q("0mm")
-        # layer_id = 1
-        # sensname = self.output_name + "_vol"
-        #
-        # for nlayer, type in zip(self.nLayers, self.typeLayers):
-        #     for ilayer in range(nlayer):
-        #         layername = self.output_name + "_stave" + "_module" + "_layer_%02i" % (layer_id)
-        #
-        #         if type == 'HG':
-        #             HGLayer_builder = self.get_builder(self.HGlayer_builder_name)
-        #             NDHPgTPCHGLayerBuilder.EndcapConfigurationLayer(HGLayer_builder, nsides, rmin, rmax, quadr, layername, sensname, "Intersection")
-        #             NDHPgTPCHGLayerBuilder.construct(HGLayer_builder, geom)
-        #             layer_lv = HGLayer_builder.get_volume(layername+"_vol")
-        #             layer_thickness = NDHPgTPCHGLayerBuilder.depth(HGLayer_builder)
-        #         if type == 'LG':
-        #             LGLayer_builder = self.get_builder(self.LGlayer_builder_name)
-        #             NDHPgTPCLGLayerBuilder.EndcapConfigurationLayer(LGLayer_builder, nsides, rmin, rmax, quadr, layername, sensname, "Intersection")
-        #             NDHPgTPCLGLayerBuilder.construct(LGLayer_builder, geom)
-        #             layer_lv = LGLayer_builder.get_volume(layername+"_vol")
-        #             layer_thickness = NDHPgTPCLGLayerBuilder.depth(LGLayer_builder)
-        #
-        #         # Placement layer in stave
-        #         layer_pos = geom.structure.Position(layername+"_pos", z=zPos + layer_thickness/2.0 - ecal_module_thickness/2.0)
-        #         layer_pla = geom.structure.Placement(layername+"_pla", volume=layer_lv, pos=layer_pos)
-        #
-        #         endcap_stave_lv.placements.append(layer_pla.name)
-        #
-        #         zPos += layer_thickness;
-        #         layer_id += 1
-        #
-        # #Place staves in the Endcap Volume
-        # module_id = -1
-        # for iend in range(2):
-        #     if iend == 0:
-        #         module_id = 0
-        #     else:
-        #         module_id = Ecal_Barrel_n_modules + 1
-        #
-        #     this_module_z_offset = (EcalEndcap_min_z + EcalEndcap_max_z)/2.
-        #     if iend == 0:
-        #         this_module_z_offset *= -1
-        #
-        #     this_module_rotY = 0.;
-        #     if iend == 0:
-        #         this_module_rotY = pi;
-        #
-        #     rotZ_offset = (pi/8. + 3.*pi/4.)
-        #
-        #     if iend == 0:
-        #         rotZ_offset = (pi/8. - pi/2.)
-        #
-        #     for iquad in range(4):
-        #         stave_id = iquad+1
-        #         this_module_rotZ = 0
-        #         if iend == 0:
-        #             this_module_rotZ = rotZ_offset - (iquad-2) * pi/2.
-        #         else:
-        #             this_module_rotZ = rotZ_offset + (iquad+1) * pi/2.
-        #
-        #         print "Placing stave ", stave_id, " and module ", module_id
-        #
-        #         #Placement staves in Endcap
-        #         name = endcap_stave_lv.name + "_%i" % (stave_id) + "_module%i" % (module_id)
-        #         endcap_stave_pos = geom.structure.Position(name + "_pos", z=this_module_z_offset )
-        #         endcap_stave_rot = geom.structure.Rotation(name + "_rot", x=Q("0deg"), y=this_module_rotY, z=this_module_rotZ+pi/4)
-        #         endcap_stave_pla = geom.structure.Placement(name + "_pla", volume=endcap_stave_lv, pos=endcap_stave_pos, rot=endcap_stave_rot)
-        #
-        #         endcap_lv.placements.append(endcap_stave_pla.name)
-        #
-        # self.add_volume(endcap_lv)
 
         # Place staves in the Endcap Volume
         sensname = self.output_name + "_vol"
@@ -701,179 +636,143 @@ class NDHPgTPCDetElementBuilder(gegede.builder.Builder):
 
         self.add_volume(endcap_lv)
 
-
-    def construct_pv(self, geom):
-        ''' construct the Pressure Vessel '''
-
-        print "Construct PV Barrel"
-
-        safety = Q("1mm")
-        ecal_module_thickness = self.get_ecal_module_thickness(geom)
-
-        nsides = self.nsides
-        pv_rInner = self.rInnerEcal + ecal_module_thickness + safety
-        pvHalfLength = self.Barrel_halfZ + ecal_module_thickness
-        half_max_dim_x = 0.5*(2 * tan( pi/nsides ) * self.rInnerEcal + ecal_module_thickness / sin( 2*pi/nsides ))
-        pv_rmin = sqrt((pv_rInner/Q("1mm"))**2 + (half_max_dim_x/Q("1mm"))**2)*Q("1mm")
-        pv_rmax = pv_rmin + self.pvThickness
-
-        # build the pressure vessel barrel
-        pvb_name = self.output_name + "Barrel"
-        pvb_shape = geom.shapes.Tubs(pvb_name, rmin=pv_rmin, rmax=pv_rmax, dz=pvHalfLength, sphi="0deg", dphi="360deg")
-        pvb_vol = geom.structure.Volume(pvb_name+"_vol", shape=pvb_shape, material=self.pvMaterial)
-
-        self.add_volume(pvb_vol)
-
-        print "Construct PV Endcap"
-
-        # build the pressure vessel endcaps
-        pv_endcap_shape_tub = geom.shapes.Tubs("PVEndcap_tubs", rmin=Q("0mm"), rmax=pv_rmax, dz=2*pvHalfLength, sphi="0deg", dphi="360deg")
-        rmin_sph = sqrt((pvHalfLength/Q("1mm"))**2 + (pv_rmax/Q("1mm"))**2)*Q("1mm")
-        rmax_sph = rmin_sph + self.pvThickness
-        pv_endcap_shape_sphere = geom.shapes.Sphere("PVEndcap_sphere", rmin=rmin_sph, rmax=rmax_sph, sphi="0deg", dphi="360deg", stheta="0deg", dtheta="179.99999deg")
-
-        pv_endcap_shape_inter = geom.shapes.Boolean( self.output_name + "Endcap", type='intersection', first=pv_endcap_shape_tub, second=pv_endcap_shape_sphere)
-        pvec_vol = geom.structure.Volume( self.output_name+"Endcap"+"_vol", shape=pv_endcap_shape_inter, material=self.pvMaterial )
-
-        self.add_volume(pvec_vol)
-
-    def construct_yoke_barrel_staves(self, geom):
-        print "Construct Yoke Barrel"
-
-        safety = Q("1mm")
-        ecal_module_thickness = self.get_ecal_module_thickness(geom)
-
-        nsides = self.nsides
-        pv_rInner = self.rInnerEcal + ecal_module_thickness + safety
-        pvHalfLength = self.Barrel_halfZ + ecal_module_thickness
-        half_max_dim_x = 0.5*(2 * tan( pi/nsides ) * self.rInnerEcal + ecal_module_thickness / sin( 2*pi/nsides ))
-        pv_rmin = sqrt((pv_rInner/Q("1mm"))**2 + (half_max_dim_x/Q("1mm"))**2)*Q("1mm")
-        pv_rmax = pv_rmin + self.pvThickness
-
-        distYoke_PV_barrel = Q("15cm")
-        yoke_rmin = pv_rmax + safety + distYoke_PV_barrel
-        yoke_rmax = yoke_rmin + self.yokeThickness
-        Yoke_Barrel_n_modules = self.nModules
-
-        rmin_sph = sqrt((pvHalfLength/Q("1mm"))**2 + (pv_rmax/Q("1mm"))**2)*Q("1mm")
-        rmax_sph = rmin_sph + self.pvThickness
-
-        dz = pvHalfLength + (rmax_sph - pvHalfLength) + distYoke_PV_barrel
-        by_name="YokeBarrel"
-
-        # sphi=self.yokePhiCutout/2.0
-        # dphi=Q("360deg")-self.yokePhiCutout
-
-        #Mother volume Barrel
-        by_shape = geom.shapes.PolyhedraRegular(by_name, numsides=nsides, rmin=yoke_rmin, rmax=yoke_rmax, dz=dz)
-        by_lv = geom.structure.Volume(by_name+"_vol", shape=by_shape, material='Air')
-
-        #Create a template module
-        yoke_stave_name = by_name + "_stave"
-        yoke_stave_volname = by_name + "_stave" + "_vol"
-
-        Yoke_Barrel_halfZ = dz
-        Yoke_Barrel_module_dim = Yoke_Barrel_halfZ * 2 / Yoke_Barrel_n_modules
-        yoke_max_dim_x = 2 * tan( pi/nsides ) * yoke_rmin + self.yokeThickness / sin( 2*pi/nsides ) #large side
-        yoke_min_dim_x = yoke_max_dim_x - 2*self.yokeThickness / tan( 2*pi/nsides ) # small side
-
-        yoke_stave_shape = geom.shapes.Trapezoid(yoke_stave_name, dx1=yoke_max_dim_x/2.0, dx2=yoke_min_dim_x/2.0,
-        dy1=Yoke_Barrel_module_dim/2.0, dy2=Yoke_Barrel_module_dim/2.0,
-        dz=self.yokeThickness/2.0)
-
-        yoke_stave_lv = geom.structure.Volume(yoke_stave_volname, shape=yoke_stave_shape, material=self.yokeMaterial)
-
-        X = yoke_rmin + self.yokeThickness / 2.
-        Y = (self.yokeThickness / 2.) / sin(2.*pi/nsides)
-        dphi = (2*pi/nsides)
-        hphi = dphi/2;
-
-        #Placing the staves and modules
-        for istave in range(nsides):
-            stave_id = istave+1
-            dstave = int( nsides/4.0 )
-            phirot =  hphi + pi/2.0
-            phirot += (istave - dstave) * dphi
-            phirot2 =  (istave - dstave) * dphi + hphi
-
-            for imodule in range(Yoke_Barrel_n_modules):
-                module_id = imodule+1
-                print "Placing stave ", stave_id, " and module ", module_id
-
-                #Placement staves in Barrel
-                name = yoke_stave_lv.name + "_%i" % (stave_id) + "_module%i" % (module_id)
-                pos = geom.structure.Position(name + "_pos", x=(X*cos(phirot2)-Y*sin(phirot2)), y=(( X*sin(phirot2)+Y*cos(phirot2) )), z=( imodule+0.5 )*Yoke_Barrel_module_dim - Yoke_Barrel_halfZ )
-                rot = geom.structure.Rotation(name + "_rot", x=pi/2.0, y=phirot+pi, z=Q("0deg"))
-                pla = geom.structure.Placement(name + "_pla", volume=yoke_stave_lv, pos=pos, rot=rot)
-
-                by_lv.placements.append(pla.name)
-
-        self.add_volume(by_lv)
-
-    def construct_yoke_endcap_staves(self, geom):
-        print "Construct Yoke Endcap"
-
-        safety = Q("1mm")
-        ecal_module_thickness = self.get_ecal_module_thickness(geom)
-
-        nsides = self.nsides
-        pv_rInner = self.rInnerEcal + ecal_module_thickness + safety
-        pvHalfLength = self.Barrel_halfZ + ecal_module_thickness
-        half_max_dim_x = 0.5*(2 * tan( pi/nsides ) * self.rInnerEcal + ecal_module_thickness / sin( 2*pi/nsides ))
-        pv_rmin = sqrt((pv_rInner/Q("1mm"))**2 + (half_max_dim_x/Q("1mm"))**2)*Q("1mm")
-        pv_rmax = pv_rmin + self.pvThickness
-
-        distYoke_PV_barrel = Q("15cm")
-        yoke_rmin = pv_rmax + safety + distYoke_PV_barrel
-        yoke_rmax = yoke_rmin + self.yokeThickness
-
-        rmin_sph = sqrt((pvHalfLength/Q("1mm"))**2 + (pv_rmax/Q("1mm"))**2)*Q("1mm")
-        rmax_sph = rmin_sph + self.pvThickness
-
-        YokeEndcap_min_z = pvHalfLength + (rmax_sph - pvHalfLength) + safety + distYoke_PV_barrel
-        YokeEndcap_max_z = pvHalfLength + (rmax_sph - pvHalfLength) + safety + distYoke_PV_barrel + self.yokeThickness
-
-        #Mother volume Endcap
-        yoke_endcap_shape_min = geom.shapes.PolyhedraRegular("YokeEndcap_min", numsides=nsides, rmin=Q("0mm"), rmax=yoke_rmax, dz=YokeEndcap_min_z)
-        yoke_endcap_shape_max = geom.shapes.PolyhedraRegular("YokeEndcap_max", numsides=nsides, rmin=Q("0mm"), rmax=yoke_rmax, dz=YokeEndcap_max_z)
-
-        yoke_endcap_shape = geom.shapes.Boolean( "YokeEndcap_sub", type='subtraction', first=yoke_endcap_shape_max, second=yoke_endcap_shape_min )
-        center_hole = geom.shapes.Tubs("center_hole", rmin=Q("0mm"), rmax=Q("0.5m"), dz=YokeEndcap_max_z, sphi=Q("0deg"), dphi=Q("360deg"))
-
-        yoke_endcap_shape_whole = geom.shapes.Boolean( "YokeEndcap", type='subtraction', first=yoke_endcap_shape, second=center_hole )
-        yoke_endcap_lv = geom.structure.Volume( self.output_name+"_vol", shape=yoke_endcap_shape_whole, material='Air' )
-
-        #Create endcap stave template
-        ecy_name = "YokeEndcap"
-        yoke_endcap_stave_name = ecy_name + "_stave"
-        yoke_endcap_stave_volname = ecy_name + "_stave" + "_vol"
-
-        yoke_endcap_stave_shape = geom.shapes.PolyhedraRegular(yoke_endcap_stave_name+"_full", numsides=nsides, rmin=Q("0mm"), rmax=yoke_rmax, dz=self.yokeThickness/2.0)
-
-        yoke_endcap_stave_shape_whole = geom.shapes.Boolean(yoke_endcap_stave_name, type='subtraction', first=yoke_endcap_stave_shape, second=center_hole )
-        yoke_endcap_stave_lv = geom.structure.Volume(yoke_endcap_stave_volname, shape=yoke_endcap_stave_shape_whole, material=self.yokeMaterial)
-
-        Yoke_Barrel_n_modules = self.nModules
-        module_id = -1
-        for iend in range(2):
-            if iend == 0:
-                module_id = 0
-            else:
-                module_id = Yoke_Barrel_n_modules + 1
-
-            this_module_z_offset = (YokeEndcap_min_z + YokeEndcap_max_z)/2.
-            if iend == 0:
-                this_module_z_offset *= -1
-
-            print "Placing yoke endcap module ", module_id
-
-            #Placement staves in Endcap
-            name = yoke_endcap_stave_lv.name + "_module%i" % (module_id)
-            yoke_endcap_stave_pos = geom.structure.Position(name + "_pos", x=Q("0mm"), y=Q("0mm"), z=this_module_z_offset)
-            #yoke_endcap_stave_rot = geom.structure.Rotation(name + "_rot", x=Q("180deg"), y=this_module_rotY, z=Q("0deg"))
-            #yoke_endcap_stave_pla = geom.structure.Placement(name + "_pla", volume=yoke_endcap_stave_lv, pos=yoke_endcap_stave_pos, rot=yoke_endcap_stave_rot)
-            yoke_endcap_stave_pla = geom.structure.Placement(name + "_pla", volume=yoke_endcap_stave_lv, pos=yoke_endcap_stave_pos)
-
-            yoke_endcap_lv.placements.append(yoke_endcap_stave_pla.name)
-
-        self.add_volume(yoke_endcap_lv)
+    # def construct_yoke_barrel_staves(self, geom):
+    #     print "Construct Yoke Barrel"
+    #
+    #     safety = Q("1mm")
+    #     ecal_module_thickness = self.get_ecal_module_thickness(geom)
+    #
+    #     nsides = self.nsides
+    #     pv_rInner = self.rInnerEcal + ecal_module_thickness + safety
+    #     pvHalfLength = self.Barrel_halfZ + ecal_module_thickness
+    #     half_max_dim_x = 0.5*(2 * tan( pi/nsides ) * self.rInnerEcal + ecal_module_thickness / sin( 2*pi/nsides ))
+    #     pv_rmin = sqrt((pv_rInner/Q("1mm"))**2 + (half_max_dim_x/Q("1mm"))**2)*Q("1mm")
+    #     pv_rmax = pv_rmin + self.pvThickness
+    #
+    #     distYoke_PV_barrel = Q("15cm")
+    #     yoke_rmin = pv_rmax + safety + distYoke_PV_barrel
+    #     yoke_rmax = yoke_rmin + self.yokeThickness
+    #     Yoke_Barrel_n_modules = self.nModules
+    #
+    #     rmin_sph = sqrt((pvHalfLength/Q("1mm"))**2 + (pv_rmax/Q("1mm"))**2)*Q("1mm")
+    #     rmax_sph = rmin_sph + self.pvThickness
+    #
+    #     dz = pvHalfLength + (rmax_sph - pvHalfLength) + distYoke_PV_barrel
+    #     by_name="YokeBarrel"
+    #
+    #     # sphi=self.yokePhiCutout/2.0
+    #     # dphi=Q("360deg")-self.yokePhiCutout
+    #
+    #     #Mother volume Barrel
+    #     by_shape = geom.shapes.PolyhedraRegular(by_name, numsides=nsides, rmin=yoke_rmin, rmax=yoke_rmax, dz=dz)
+    #     by_lv = geom.structure.Volume(by_name+"_vol", shape=by_shape, material='Air')
+    #
+    #     #Create a template module
+    #     yoke_stave_name = by_name + "_stave"
+    #     yoke_stave_volname = by_name + "_stave" + "_vol"
+    #
+    #     Yoke_Barrel_halfZ = dz
+    #     Yoke_Barrel_module_dim = Yoke_Barrel_halfZ * 2 / Yoke_Barrel_n_modules
+    #     yoke_max_dim_x = 2 * tan( pi/nsides ) * yoke_rmin + self.yokeThickness / sin( 2*pi/nsides ) #large side
+    #     yoke_min_dim_x = yoke_max_dim_x - 2*self.yokeThickness / tan( 2*pi/nsides ) # small side
+    #
+    #     yoke_stave_shape = geom.shapes.Trapezoid(yoke_stave_name, dx1=yoke_max_dim_x/2.0, dx2=yoke_min_dim_x/2.0,
+    #     dy1=Yoke_Barrel_module_dim/2.0, dy2=Yoke_Barrel_module_dim/2.0,
+    #     dz=self.yokeThickness/2.0)
+    #
+    #     yoke_stave_lv = geom.structure.Volume(yoke_stave_volname, shape=yoke_stave_shape, material=self.yokeMaterial)
+    #
+    #     X = yoke_rmin + self.yokeThickness / 2.
+    #     Y = (self.yokeThickness / 2.) / sin(2.*pi/nsides)
+    #     dphi = (2*pi/nsides)
+    #     hphi = dphi/2;
+    #
+    #     #Placing the staves and modules
+    #     for istave in range(nsides):
+    #         stave_id = istave+1
+    #         dstave = int( nsides/4.0 )
+    #         phirot =  hphi + pi/2.0
+    #         phirot += (istave - dstave) * dphi
+    #         phirot2 =  (istave - dstave) * dphi + hphi
+    #
+    #         for imodule in range(Yoke_Barrel_n_modules):
+    #             module_id = imodule+1
+    #             print "Placing stave ", stave_id, " and module ", module_id
+    #
+    #             #Placement staves in Barrel
+    #             name = yoke_stave_lv.name + "_%i" % (stave_id) + "_module%i" % (module_id)
+    #             pos = geom.structure.Position(name + "_pos", x=(X*cos(phirot2)-Y*sin(phirot2)), y=(( X*sin(phirot2)+Y*cos(phirot2) )), z=( imodule+0.5 )*Yoke_Barrel_module_dim - Yoke_Barrel_halfZ )
+    #             rot = geom.structure.Rotation(name + "_rot", x=pi/2.0, y=phirot+pi, z=Q("0deg"))
+    #             pla = geom.structure.Placement(name + "_pla", volume=yoke_stave_lv, pos=pos, rot=rot)
+    #
+    #             by_lv.placements.append(pla.name)
+    #
+    #     self.add_volume(by_lv)
+    #
+    # def construct_yoke_endcap_staves(self, geom):
+    #     print "Construct Yoke Endcap"
+    #
+    #     safety = Q("1mm")
+    #     ecal_module_thickness = self.get_ecal_module_thickness(geom)
+    #
+    #     nsides = self.nsides
+    #     pv_rInner = self.rInnerEcal + ecal_module_thickness + safety
+    #     pvHalfLength = self.Barrel_halfZ + ecal_module_thickness
+    #     half_max_dim_x = 0.5*(2 * tan( pi/nsides ) * self.rInnerEcal + ecal_module_thickness / sin( 2*pi/nsides ))
+    #     pv_rmin = sqrt((pv_rInner/Q("1mm"))**2 + (half_max_dim_x/Q("1mm"))**2)*Q("1mm")
+    #     pv_rmax = pv_rmin + self.pvThickness
+    #
+    #     distYoke_PV_barrel = Q("15cm")
+    #     yoke_rmin = pv_rmax + safety + distYoke_PV_barrel
+    #     yoke_rmax = yoke_rmin + self.yokeThickness
+    #
+    #     rmin_sph = sqrt((pvHalfLength/Q("1mm"))**2 + (pv_rmax/Q("1mm"))**2)*Q("1mm")
+    #     rmax_sph = rmin_sph + self.pvThickness
+    #
+    #     YokeEndcap_min_z = pvHalfLength + (rmax_sph - pvHalfLength) + safety + distYoke_PV_barrel
+    #     YokeEndcap_max_z = pvHalfLength + (rmax_sph - pvHalfLength) + safety + distYoke_PV_barrel + self.yokeThickness
+    #
+    #     #Mother volume Endcap
+    #     yoke_endcap_shape_min = geom.shapes.PolyhedraRegular("YokeEndcap_min", numsides=nsides, rmin=Q("0mm"), rmax=yoke_rmax, dz=YokeEndcap_min_z)
+    #     yoke_endcap_shape_max = geom.shapes.PolyhedraRegular("YokeEndcap_max", numsides=nsides, rmin=Q("0mm"), rmax=yoke_rmax, dz=YokeEndcap_max_z)
+    #
+    #     yoke_endcap_shape = geom.shapes.Boolean( "YokeEndcap_sub", type='subtraction', first=yoke_endcap_shape_max, second=yoke_endcap_shape_min )
+    #     center_hole = geom.shapes.Tubs("center_hole", rmin=Q("0mm"), rmax=Q("0.5m"), dz=YokeEndcap_max_z, sphi=Q("0deg"), dphi=Q("360deg"))
+    #
+    #     yoke_endcap_shape_whole = geom.shapes.Boolean( "YokeEndcap", type='subtraction', first=yoke_endcap_shape, second=center_hole )
+    #     yoke_endcap_lv = geom.structure.Volume( self.output_name+"_vol", shape=yoke_endcap_shape_whole, material='Air' )
+    #
+    #     #Create endcap stave template
+    #     ecy_name = "YokeEndcap"
+    #     yoke_endcap_stave_name = ecy_name + "_stave"
+    #     yoke_endcap_stave_volname = ecy_name + "_stave" + "_vol"
+    #
+    #     yoke_endcap_stave_shape = geom.shapes.PolyhedraRegular(yoke_endcap_stave_name+"_full", numsides=nsides, rmin=Q("0mm"), rmax=yoke_rmax, dz=self.yokeThickness/2.0)
+    #
+    #     yoke_endcap_stave_shape_whole = geom.shapes.Boolean(yoke_endcap_stave_name, type='subtraction', first=yoke_endcap_stave_shape, second=center_hole )
+    #     yoke_endcap_stave_lv = geom.structure.Volume(yoke_endcap_stave_volname, shape=yoke_endcap_stave_shape_whole, material=self.yokeMaterial)
+    #
+    #     Yoke_Barrel_n_modules = self.nModules
+    #     module_id = -1
+    #     for iend in range(2):
+    #         if iend == 0:
+    #             module_id = 0
+    #         else:
+    #             module_id = Yoke_Barrel_n_modules + 1
+    #
+    #         this_module_z_offset = (YokeEndcap_min_z + YokeEndcap_max_z)/2.
+    #         if iend == 0:
+    #             this_module_z_offset *= -1
+    #
+    #         print "Placing yoke endcap module ", module_id
+    #
+    #         #Placement staves in Endcap
+    #         name = yoke_endcap_stave_lv.name + "_module%i" % (module_id)
+    #         yoke_endcap_stave_pos = geom.structure.Position(name + "_pos", x=Q("0mm"), y=Q("0mm"), z=this_module_z_offset)
+    #         #yoke_endcap_stave_rot = geom.structure.Rotation(name + "_rot", x=Q("180deg"), y=this_module_rotY, z=Q("0deg"))
+    #         #yoke_endcap_stave_pla = geom.structure.Placement(name + "_pla", volume=yoke_endcap_stave_lv, pos=yoke_endcap_stave_pos, rot=yoke_endcap_stave_rot)
+    #         yoke_endcap_stave_pla = geom.structure.Placement(name + "_pla", volume=yoke_endcap_stave_lv, pos=yoke_endcap_stave_pos)
+    #
+    #         yoke_endcap_lv.placements.append(yoke_endcap_stave_pla.name)
+    #
+    #     self.add_volume(yoke_endcap_lv)
