--- duneggd/Config/CylindricalMPT.cfg
@@ -1,156 +1,0 @@
-# run like
-# gegede-cli duneggd/Config/CylindricalMPT.cfg \
-# duneggd/Config/DETENCLOSURE_MPT_Test.cfg \
-# duneggd/Config/WORLDggd.cfg -w World -o mpt.gdml
-
-[CylindricalMPT]
-class               = duneggd.SubDetector.CylindricalMPT.CylindricalMPTBuilder
-subbuilders         = ['GArTPC','InnerBarrelECalBuilder',"InnerEndcapECalBuilder","OuterBarrelECalBuilder"]
-buildGarTPC    	    = True
-yokeMaterial="Iron"
-yokeInnerR=Q("3.20m")
-yokeInnerZ=Q("3.9m")
-yokeThicknessR=Q("0.5m")
-yokeThicknessZ=Q("0.5m")
-yokeBufferToBoundaryR=Q("0.0m")
-yokeBufferToBoundaryZ=Q("0.0m")
-yokePhiCutout=Q("90deg")
-buildYoke=True
-innerBField="0.4 T, 0.0 T, 0.0 T"
-buildGarTPC=True
-buildInnerEcal=True
-buildOuterEcal=True
-
-# starting location of the ECalEndcaps
-# needs to agree with chamber dimension dz
-# issue is that this requires asking the GArTPC builder
-# should try to refactorize to fix this dependency
-IBECalXStart=Q("260cm")
-#
-pvInnerRadius=Q("285cm")
-pvThickness=Q("3cm")
-pvHalfLength=Q("285cm")
-pvEndCapBulge=Q("100cm")
-pvMaterial='Steel'
-buildPV=True
-###############################
-# GArTPC
-###############################
-[GArTPC]
-class               = duneggd.SubDetector.GArTPC.GArTPCBuilder
-subbuilders         = []
-# this builds an active TPC out to a radius of 260cm (tpcDimension)
-# and length 250cm*2=500cm
-# It builds a gas buffer from 260cm out to a radius of 274cm
-# from 250cm to 520cm in length
-#
-#   1) The buffer is called TPCGasVol out to a radius of 2.73m
-#      and a length of 2.58m
-#
-#   2) The buffer is called TPCChamber the last 1cm in each direction.
-#      The original idea was that the TPChamber volume was the steel
-#      pressure vessel. Now we want to stick an ECAL inside the vessel,
-#      so here we just make it more gas and we'll make our own vessel
-#      at a larger radius.
-#
-halfDimension       = {'rmin':Q('0cm'),'rmax':Q('274cm'),'dz':Q('260cm')}
-tpcDimension        = {'r':Q('260cm'),'dz':Q('250cm')}
-chamberDimension    = { 'r':Q('274cm'),'dz':Q('520cm') }
-ChamberMaterial = 'HP_ArCO2'
-GasType            = 'HP_ArCO2'
-drift               = 'z'
-
-
-############################
-# InnerBarrelECal
-############################
-[IBTileBuilder]
-class		    = duneggd.Active.MPTECalComponents.MPTECalTileBuilder
-subbuilders	    = []
-dx = Q("20mm")
-dy = Q("20mm")
-mat=['Lead', 'Scintillator', 'FR4']
-output_name="IBTile"
-
-[IBStripBuilder]
-class		    = duneggd.Active.MPTECalComponents.MPTECalStripBuilder
-tile_builder_name='IBTileBuilder'
-subbuilders	    = [tile_builder_name]
-extra_space = Q("0.1mm")
-length=Q("520cm")
-output_name="IBStrip"
-
-[InnerBarrelECalBuilder]
-class		    = duneggd.Active.MPTECalComponents.MPTECalLayerBuilder
-strip_builder_name='IBStripBuilder'
-tile_builder_name='IBTileBuilder'
-subbuilders	    = [strip_builder_name,tile_builder_name]
-extra_space = Q("0.1mm")
-layer_gap = Q("0.0mm")
-r=Q("274cm")
-geometry = 'cylinder'
-nlayers = 10
-output_name = "InnerBarrelECal"
-
-############################
-# InnerEndcapECal
-# z location set to begin just outside the end of the chamber
-############################
-[IECTileBuilder]
-class		    = duneggd.Active.MPTECalComponents.MPTECalTileBuilder
-subbuilders	    = []
-dx = Q("20mm")
-dy = Q("20mm")
-mat=['Lead', 'Scintillator', 'FR4']
-output_name="IECTile"
-
-[IECStripBuilder]
-class		    = duneggd.Active.MPTECalComponents.MPTECalStripBuilder
-tile_builder_name='IECTileBuilder'
-subbuilders	    = [tile_builder_name]
-extra_space = Q("0.1mm")
-length=Q("560cm")
-output_name="IECStrip"
-
-[InnerEndcapECalBuilder]
-class		    = duneggd.Active.MPTECalComponents.MPTECalLayerBuilder
-strip_builder_name='IECStripBuilder'
-tile_builder_name='IECTileBuilder'
-subbuilders	    = [strip_builder_name,tile_builder_name]
-extra_space = Q("0.1mm")
-layer_gap = Q("0.0mm")
-r=Q("274cm")
-geometry = 'cplane'
-nlayers = 25
-output_name = "InnerEndcapECal"
-
-############################
-# OuterBarrelECal
-############################
-[OBTileBuilder]
-class		    = duneggd.Active.MPTECalComponents.MPTECalTileBuilder
-subbuilders	    = []
-dx = Q("20mm")
-dy = Q("20mm")
-mat=['Lead', 'Scintillator', 'FR4']
-output_name="OBTile"
-
-[OBStripBuilder]
-class		    = duneggd.Active.MPTECalComponents.MPTECalStripBuilder
-tile_builder_name='OBTileBuilder'
-subbuilders	    = [tile_builder_name]
-extra_space = Q("0.1mm")
-length=Q("600cm")
-output_name="OBStrip"
-
-[OuterBarrelECalBuilder]
-class		    = duneggd.Active.MPTECalComponents.MPTECalLayerBuilder
-strip_builder_name='OBStripBuilder'
-tile_builder_name='OBTileBuilder'
-subbuilders	    = [strip_builder_name,tile_builder_name]
-extra_space = Q("0.1mm")
-layer_gap = Q("0.0mm")
-r=Q("289cm")
-geometry = 'cylinder'
-nlayers = 20
-output_name = "OuterBarrelECal"
