+++ duneggd/Config/DETENCLOSURE_ArgonCube_MPD.cfg
@@ -1,0 +1,14 @@
+# gegede-cli duneggd/Config/MPD_Concept.cfg \
+# duneggd/Config/ArgonCube/*.cfg \
+# duneggd/Config/DETENCLOSURE_ArgonCube_MPD.cfg \
+# duneggd/Config/WORLDggd.cfg -w World -o ArgonCube_MPD.gdml
+
+[DetEnclosure]
+class               = duneggd.DetEnc.DetEncBuilder
+subbuilders         = ['ArgonCubeDetector','MPD']
+# Mike Kordosky: dimension choosen here to exactly match
+# the hall used by the ND task force  (site.xml)
+halfDimension       = {'dx':Q('899.16cm'),'dy':Q('768.096cm'),'dz':Q('1501.61625cm')}
+Material            = 'Air'
+Positions	    = [	[ Q("0m"), Q("-146.8cm"), Q("-425.6cm") ], [ Q("0m"), Q("-214.9cm"), Q("384.4cm") ] ]
+Rotations	    = [	[ Q("0deg"), Q("0deg"), Q("0deg") ], [ Q("0deg"), Q("90deg"), Q("0deg") ] ]
