+++ duneggd/Config/DETENCLOSURE_MPD.cfg
@@ -1,0 +1,13 @@
+# gegede-cli  duneggd/Config/MPD_Concept.cfg \
+# duneggd/Config/DETENCLOSURE_MPD.cfg \
+# duneggd/Config/WORLDggd.cfg -w World -o MPD.gdml
+
+[DetEnclosure]
+class               = duneggd.DetEnc.DetEncBuilder
+subbuilders         = ['MPD']
+# Mike Kordosky: dimension choosen here to exactly match
+# the hall used by the ND task force  (site.xml)
+halfDimension       = {'dx':Q('899.16cm'),'dy':Q('768.096cm'),'dz':Q('1501.61625cm')}
+Material            = 'Air'
+Positions	    = [	[ Q("0m"), Q("0m"), Q("0m") ]]
+Rotations	    = [	[ Q("0deg"), Q("90deg"), Q("0deg") ]]
