--- duneggd/Config/DETENCLOSURE_MPT_Test.cfg
@@ -1,8 +1,0 @@
-[DetEnclosure]
-class               = duneggd.DetEnc.DetEncBuilder
-subbuilders         = ['CylindricalMPT']
-halfDimension       = {'dx':Q('8m'),'dy':Q('8m'),'dz':Q('10m')}
-Material            = 'Air'
-Positions	    = [	[ Q("0m"), Q("0m"), Q("0m") ]]
-Rotations	    = [	[ Q("0deg"), Q("0deg"), Q("0deg") ]]
-
