--- duneggd/Config/DETENCLOSURE_MPT_concept.cfg
@@ -1,25 +1,0 @@
-# How to run, from the dunendggd directory:
-# gegede-cli  duneggd/Config/CylindricalMPT.cfg \
-# duneggd/Config/ArgonCube/ArgonCubeDetector.cfg \
-# duneggd/Config/DETENCLOSURE_MPT_concept.cfg \
-# duneggd/Config/ArgonCube/*.cfg duneggd/Config/WORLDggd.cfg -w World -o lar_mpt.gdml
-
-[DetEnclosure]
-class               = duneggd.DetEnc.DetEncBuilder
-Material            = 'Air'
-# Mike Kordosky: dimension choosen here to exactly match
-# the hall used by the ND task force  (site.xml)
-halfDimension       = {'dx':Q('899.16cm'),'dy':Q('768.096cm'),'dz':Q('1501.61625cm')}
-
-####Comment in or out the following sections to switch from IronDipole to KLOE
-######## Argon Cube + IronDipole
-subbuilders         = ['ArgonCubeDetector','CylindricalMPT']
-
-### v5 positions (move to agree with NDTF beam in hall)
-Positions	    = [	[ Q("0m"), Q("-200.0cm"), Q("-480.3cm") ],
-	              [ Q("0m"), Q("-198.3cm"), Q("220.4cm") ] ,
-	              [ Q("0m"), Q("0m"), Q("0m") ] ]
-
-Rotations	    = [	[ Q("0deg"), Q("0deg"), Q("0deg") ],
-		      	[ Q("0deg"), Q("0deg"), Q("0deg") ] ,
-		      	[ Q("0deg"), Q("0deg"), Q("0deg") ] ]
