+++ duneggd/Config/MPD_Concept.cfg
@@ -1,20 +1,8 @@
 # run like
-# gegede-cli NDHPgTPC_v02.cfg duneggd/Config/ArgonCube/*.cfg duneggd/Config/WORLDggd.cfg -w World -o NDHPgTPC_v02.gdml
+# gegede-cli MPD_Concept.cfg -o MPD_Concept.gdml
 
-[DetEnclosure]
-class               = duneggd.DetEnc.DetEncBuilder
-subbuilders         = ['ArgonCubeDetector','NDHPgTPC']
-halfDimension       = {'dx':Q('899.16cm'),'dy':Q('768.096cm'),'dz':Q('1501.61625cm')}
-Material            = 'Air'
-Positions	    = [	[ Q("0m"), Q("-146.8cm"), Q("-425.6cm") ],
-	              [ Q("0m"), Q("-214.9cm"), Q("384.4cm") ] ]
-
-Rotations	    = [	[ Q("0deg"), Q("0deg"), Q("0deg") ],
-		      	[ Q("0deg"), Q("90deg"), Q("0deg") ] ]
-
-
-[NDHPgTPC]
-class               = duneggd.SubDetector.NDHPgTPC_v03.NDHPgTPC_v03_Builder
+[MPD]
+class               = duneggd.SubDetector.NDHPgTPC.NDHPgTPC_Builder
 subbuilders         = ['GArTPC','ECALBarrelBuilder','ECALEndcapBuilder','PVBuilder', 'MagnetBuilder']
 #halfDimension       = {'dx':Q('5m'),'dy':Q('5m'),'dz':Q('5m')}
 
@@ -54,7 +42,7 @@ GasType             = 'HP_ArCH4'
 drift               = 'z'
 
 [HGLayerBuilder1]
-class		    = duneggd.Active.NDHPgTPC_v03_Components.NDHPgTPCHGLayerBuilder
+class		    = duneggd.Active.NDHPgTPC.NDHPgTPCHGLayerBuilder
 subbuilders	    = []
 dx = Q("100mm")
 dy = Q("100mm")
@@ -64,7 +52,7 @@ mat=['Copper', 'Scintillator', 'FR4']
 output_name="MPTECalHGLayerType1"
 
 [HGLayerBuilder2]
-class		    = duneggd.Active.NDHPgTPC_v03_Components.NDHPgTPCHGLayerBuilder
+class		    = duneggd.Active.NDHPgTPC.NDHPgTPCHGLayerBuilder
 subbuilders	    = []
 dx = Q("100mm")
 dy = Q("100mm")
@@ -74,7 +62,7 @@ mat=['Copper', 'Scintillator', 'FR4']
 output_name="MPTECalHGLayerType2"
 
 [LGLayerBuilder1]
-class		    = duneggd.Active.NDHPgTPC_v03_Components.NDHPgTPCLGLayerBuilder
+class		    = duneggd.Active.NDHPgTPC.NDHPgTPCLGLayerBuilder
 subbuilders	    = []
 dx = Q("100mm")
 dy = Q("100mm")
@@ -84,7 +72,7 @@ mat=['Copper', 'Scintillator']
 output_name="MPTECalLGLayerType1"
 
 [LGLayerBuilder2]
-class		    = duneggd.Active.NDHPgTPC_v03_Components.NDHPgTPCLGLayerBuilder
+class		    = duneggd.Active.NDHPgTPC.NDHPgTPCLGLayerBuilder
 subbuilders	    = []
 dx = Q("100mm")
 dy = Q("100mm")
@@ -98,7 +86,7 @@ output_name="MPTECalLGLayerType2"
 ############################
 
 [ECALBarrelBuilder]
-class		    = duneggd.Active.NDHPgTPC_v03_Components.NDHPgTPCDetElementBuilder
+class		    = duneggd.Active.NDHPgTPC.NDHPgTPCDetElementBuilder
 HGlayerType1_builder_name='HGLayerBuilder1'
 HGlayerType2_builder_name='HGLayerBuilder2'
 LGlayerType1_builder_name='LGLayerBuilder1'
@@ -120,7 +108,7 @@ typeLayers=['HG1', 'LG1']
 ############################
 
 [ECALEndcapBuilder]
-class		    = duneggd.Active.NDHPgTPC_v03_Components.NDHPgTPCDetElementBuilder
+class		    = duneggd.Active.NDHPgTPC.NDHPgTPCDetElementBuilder
 HGlayerType1_builder_name='HGLayerBuilder1'
 HGlayerType2_builder_name='HGLayerBuilder2'
 LGlayerType1_builder_name='LGLayerBuilder1'
@@ -142,7 +130,7 @@ typeLayers=['HG1', 'LG1']
 ############################
 
 [PVBuilder]
-class		    = duneggd.Active.NDHPgTPC_v03_Components.NDHPgTPCDetElementBuilder
+class		    = duneggd.Active.NDHPgTPC.NDHPgTPCDetElementBuilder
 HGlayerType1_builder_name='HGLayerBuilder1'
 HGlayerType2_builder_name='HGLayerBuilder2'
 LGlayerType1_builder_name='LGLayerBuilder1'
@@ -162,7 +150,7 @@ typeLayers=['HG1', 'LG1']
 ############################
 
 [MagnetBuilder]
-class		    = duneggd.Active.NDHPgTPC_v03_Components.NDHPgTPCDetElementBuilder
+class		    = duneggd.Active.NDHPgTPC.NDHPgTPCDetElementBuilder
 HGlayerType1_builder_name='HGLayerBuilder1'
 HGlayerType2_builder_name='HGLayerBuilder2'
 LGlayerType1_builder_name='LGLayerBuilder1'
