--- duneggd/Config/ND_Concept_all.cfg
@@ -1,157 +1,0 @@
-# run like
-# gegede-cli ND_Concept_all.cfg \
-# ../../dunendggd/duneggd/Config/ArgonCube/*.cfg \
-# ../../dunendggd/duneggd/Config/WORLDggd.cfg -w World -o ND_Concept_all.gdml
-
-[DetEnclosure]
-class               = duneggd.DetEnc.DetEncBuilder
-subbuilders         = ['ArgonCubeDetector','NDHPgTPC']
-halfDimension       = {'dx':Q('899.16cm'),'dy':Q('768.096cm'),'dz':Q('1501.61625cm')}
-Material            = 'Air'
-Positions	    = [	[ Q("0m"), Q("-146.8cm"), Q("-425.6cm") ],
-	              [ Q("0m"), Q("-214.9cm"), Q("384.4cm") ] ]
-
-Rotations	    = [	[ Q("0deg"), Q("0deg"), Q("0deg") ],
-		      	[ Q("0deg"), Q("0deg"), Q("0deg") ] ]
-
-[NDHPgTPC]
-class               = duneggd.SubDetector.NDHPgTPCDetector.NDHPgTPCDetectorBuilder
-#subbuilders         = ['GArTPC','InnerBarrelECalBuilder', 'InnerEndcapECalBuilder','OuterBarrelECalBuilder']
-subbuilders         = ['GArTPC','InnerEndcapECalBuilder','InnerBarrelECalBuilder','OuterBarrelECalBuilder']
-
-## Parameters
-yokeMaterial="Iron"
-yokeInnerR=Q("3390mm")
-yokeInnerZ=Q("4270mm")
-yokeThicknessR=Q("500mm")
-yokeThicknessZ=Q("500mm")
-yokeBufferToBoundaryR=Q("0.0m")
-yokeBufferToBoundaryZ=Q("0.0m")
-yokePhiCutout=Q("90deg")
-buildYoke=True
-innerBField="0.4 T, 0.0 T, 0.0 T"
-buildGarTPC=True
-buildOuterEcal=True
-buildInnerEcalEndcap=True
-buildInnerEcalBarrel=True
-
-# starting location of the ECalEndcaps
-# needs to agree with chamber dimension dz
-# issue is that this requires asking the GArTPC builder
-# should try to refactorize to fix this dependency
-IBECalXStart=Q("2600mm")
-#
-pvInnerRadius=Q("2980mm")
-pvThickness=Q("20mm")
-pvHalfLength=Q("3240mm")
-pvEndCapBulge=Q("1000mm")
-pvMaterial='Steel'
-buildPV=True
-
-###############################
-# GArTPC
-###############################
-[GArTPC]
-class               = duneggd.SubDetector.GArTPC.GArTPCBuilder
-subbuilders         = []
-# this builds an active TPC out to a radius of 260cm (tpcDimension)
-# and length 250cm*2=500cm
-# It builds a gas buffer from 260cm out to a radius of 274cm
-# from 250cm to 520cm in length
-#
-#   1) The buffer is called TPCGasVol out to a radius of 2.73m
-#      and a length of 2.58m
-#
-#   2) The buffer is called TPCChamber the last 1cm in each direction.
-#      The original idea was that the TPChamber volume was the steel
-#      pressure vessel. Now we want to stick an ECAL inside the vessel,
-#      so here we just make it more gas and we'll make our own vessel
-#      at a larger radius.
-#
-halfDimension       = {'rmin':Q('0mm'),'rmax':Q('2740mm'),'dz':Q('2600mm')}
-tpcDimension        = {'r':Q('2600mm'),'dz':Q('2500mm')}
-chamberDimension    = { 'r':Q('2740mm'),'dz':Q('5200mm') }
-ChamberMaterial     = 'HP_ArCH4'
-GasType             = 'HP_ArCH4'
-drift               = 'z'
-
-[IBStripBuilder]
-class		    = duneggd.Active.NDHPgTPC.NDHPgTPCStripBuilder
-subbuilders	    = []
-dx = Q("20mm")
-dy = Q("20mm")
-dz = [Q("2mm"),Q("5mm"),Q("1mm")]
-lspacing=[Q('0mm'), Q('0mm'), Q('0mm')]
-mat=['Copper', 'Scintillator', 'FR4']
-output_name="IBStrip"
-length=Q("6480mm")
-extra_space = Q("0.0mm")
-
-[OBStripBuilder]
-class		    = duneggd.Active.NDHPgTPC.NDHPgTPCStripBuilder
-subbuilders	    = []
-dx = Q("20mm")
-dy = Q("20mm")
-dz = [Q("2mm"),Q("5mm"),Q("1mm")]
-lspacing=[Q('0mm'), Q('0mm'), Q('0mm')]
-mat=['Copper', 'Scintillator', 'FR4']
-extra_space = Q("0.0mm")
-length=Q("6480mm")
-output_name="OBStrip"
-
-[IECLayerBuilder]
-class		    = duneggd.Active.NDHPgTPC.NDHPgTPCEndcapLayerBuilder
-subbuilders	    = []
-r=Q("2740mm")
-dz = [Q("2mm"),Q("5mm"),Q("1mm")]
-lspacing=[Q('0mm'), Q('0mm'), Q('0mm')]
-mat=['Copper', 'Scintillator', 'FR4']
-output_name="IECLayer"
-extra_space = Q("0.0mm")
-
-############################
-# InnerBarrelECal
-############################
-
-[InnerBarrelECalBuilder]
-class		    = duneggd.Active.NDHPgTPC.NDHPgTPCLayerBuilder
-strip_builder_name='IBStripBuilder'
-subbuilders	    = [strip_builder_name]
-extra_space = Q("0.0mm")
-layer_gap = Q("0.0mm")
-r=Q("2740mm")
-geometry = 'cylinder'
-nlayers = 30
-output_name = "InnerBarrelECal"
-
-############################
-# OuterBarrelECal
-############################
-
-[OuterBarrelECalBuilder]
-class		    = duneggd.Active.NDHPgTPC.NDHPgTPCLayerBuilder
-strip_builder_name='OBStripBuilder'
-subbuilders	    = [strip_builder_name]
-extra_space = Q("0.0mm")
-layer_gap = Q("0.0mm")
-r=Q("3000mm")
-geometry = 'cylinder'
-nlayers = 50
-output_name = "OuterBarrelECal"
-
-############################
-# InnerEndcapECal
-# z location set to begin just outside the end of the chamber
-############################
-
-[InnerEndcapECalBuilder]
-class		    = duneggd.Active.NDHPgTPC.NDHPgTPCLayerBuilder
-endcap_layer_builder_name='IECLayerBuilder'
-strip_builder_name='IBStripBuilder'
-subbuilders	    = [endcap_layer_builder_name,strip_builder_name]
-extra_space = Q("0.0mm")
-layer_gap = Q("0.0mm")
-r=Q("2740mm")
-geometry = 'cplane'
-nlayers = 80
-output_name = "InnerEndcapECal"
