--- duneggd/Config/ND_Strawman_Concept_v01.cfg
@@ -1,167 +1,0 @@
-# run like
-# gegede-cli ND_Strawman_Concept_v01.cfg -w World -o ND_Strawman_Concept_v01.gdml
-
-[World]
-class               = duneggd.World.WorldBuilder
-subbuilders         = ['NDHPgTPC']
-halfDimension       = {'dx':Q('10m'),'dy':Q('10m'),'dz':Q('10m')}
-DetEncPosition       = [Q('0m'),Q('0m'),Q('0m')]
-DetEncRotation       = [Q('0deg'),Q('90deg'),Q('0deg')]
-Material            = 'Air'
-
-[NDHPgTPC]
-class               = duneggd.SubDetector.NDHPgTPC_v02.NDHPgTPC_v02_Builder
-subbuilders         = ['GArTPC','ECALBarrelBuilder','ECALEndcapBuilder','PVBuilder','YokeBarrelBuilder','YokeEndcapBuilder']
-halfDimension       = {'dx':Q('10m'),'dy':Q('10m'),'dz':Q('10m')}
-
-## Parameters
-buildYoke=True
-innerBField="0.4 T, 0.0 T, 0.0 T"
-buildGarTPC=True
-buildEcal=True
-buildPV=True
-
-###############################
-# GArTPC
-###############################
-[GArTPC]
-class               = duneggd.SubDetector.GArTPC.GArTPCBuilder
-subbuilders         = []
-# this builds an active TPC out to a radius of 260cm (tpcDimension)
-# and length 250cm*2=500cm
-# It builds a gas buffer from 260cm out to a radius of 274cm
-# from 250cm to 520cm in length
-#
-#   1) The buffer is called TPCGasVol out to a radius of 2.73m
-#      and a length of 2.58m
-#
-#   2) The buffer is called TPCChamber the last 1cm in each direction.
-#      The original idea was that the TPChamber volume was the steel
-#      pressure vessel. Now we want to stick an ECAL inside the vessel,
-#      so here we just make it more gas and we'll make our own vessel
-#      at a larger radius.
-#
-halfDimension       = {'rmin':Q('0mm'),'rmax':Q('2740mm'),'dz':Q('2600mm')}
-tpcDimension        = {'r':Q('2600mm'),'dz':Q('2500mm')}
-chamberDimension    = { 'r':Q('2740mm'),'dz':Q('5200mm') }
-ChamberMaterial     = 'HP_ArCH4'
-GasType             = 'HP_ArCH4'
-drift               = 'z'
-
-[HGLayerBuilder]
-class		    = duneggd.Active.NDHPgTPC_v02_Components.NDHPgTPCHGLayerBuilder
-subbuilders	    = []
-dx = Q("100mm")
-dy = Q("100mm")
-dz = [Q("2mm"),Q("10mm"),Q("1mm")]
-lspacing=[Q('0.1mm'), Q('0.1mm'), Q('0.1mm')]
-mat=['Copper', 'Scintillator', 'FR4']
-output_name="MPTECalHGLayer"
-
-[LGLayerBuilder]
-class		    = duneggd.Active.NDHPgTPC_v02_Components.NDHPgTPCLGLayerBuilder
-subbuilders	    = []
-dx = Q("100mm")
-dy = Q("100mm")
-dz = [Q("4mm"),Q("5mm"),Q("5mm")]
-lspacing=[Q('0.1mm'), Q('0.1mm'), Q('0.1mm')]
-mat=['Copper', 'Scintillator', 'Scintillator']
-output_name="MPTECalLGLayer"
-
-############################
-# ECAL Barrel
-############################
-
-[ECALBarrelBuilder]
-class		    = duneggd.Active.NDHPgTPC_v02_Components.NDHPgTPCDetElementBuilder
-HGlayer_builder_name='HGLayerBuilder'
-LGlayer_builder_name='LGLayerBuilder'
-subbuilders	    = [HGlayer_builder_name,LGlayer_builder_name]
-geometry = 'ECALBarrel'
-nsides = 8
-nModules = 5
-output_name = "BarrelECal"
-rInnerEcal=Q("2740mm")
-Barrel_halfZ = Q('2600mm')
-nLayers=[4, 16, 4, 31]
-typeLayers=['HG', 'LG', 'HG', 'LG']
-
-############################
-# ECAL Endcap
-############################
-
-[ECALEndcapBuilder]
-class		    = duneggd.Active.NDHPgTPC_v02_Components.NDHPgTPCDetElementBuilder
-HGlayer_builder_name='HGLayerBuilder'
-LGlayer_builder_name='LGLayerBuilder'
-subbuilders	    = [HGlayer_builder_name,LGlayer_builder_name]
-geometry = 'ECALEndcap'
-nsides = 8
-nModules = 5
-output_name = "EndcapECal"
-rInnerEcal=Q("2740mm")
-Barrel_halfZ = Q('2600mm')
-nLayers=[4, 16, 2, 23]
-typeLayers=['HG', 'LG', 'HG', 'LG']
-
-############################
-# Pressure Vessel
-############################
-
-[PVBuilder]
-class		    = duneggd.Active.NDHPgTPC_v02_Components.NDHPgTPCDetElementBuilder
-HGlayer_builder_name='HGLayerBuilder'
-LGlayer_builder_name='LGLayerBuilder'
-subbuilders	    = [HGlayer_builder_name,LGlayer_builder_name]
-geometry = 'PV'
-output_name = "PV"
-pvMaterial = "Steel"
-pvThickness = Q("20mm")
-rInnerEcal=Q("2740mm")
-Barrel_halfZ = Q('2600mm')
-nLayers=[4, 16, 4, 31]
-typeLayers=['HG', 'LG', 'HG', 'LG']
-
-############################
-# Yoke Barrel
-############################
-
-[YokeBarrelBuilder]
-class		    = duneggd.Active.NDHPgTPC_v02_Components.NDHPgTPCDetElementBuilder
-HGlayer_builder_name='HGLayerBuilder'
-LGlayer_builder_name='LGLayerBuilder'
-subbuilders	    = [HGlayer_builder_name,LGlayer_builder_name]
-geometry = 'YokeBarrel'
-nsides = 12
-nModules = 5
-output_name = "BarrelYoke"
-pvThickness = Q("20mm")
-yokePhiCutout = Q("90deg")
-yokeThickness = Q("500mm")
-yokeMaterial = "Steel"
-rInnerEcal=Q("2740mm")
-Barrel_halfZ = Q('2600mm')
-nLayers=[4, 16, 4, 31]
-typeLayers=['HG', 'LG', 'HG', 'LG']
-
-############################
-# Yoke Endcap
-############################
-
-[YokeEndcapBuilder]
-class		    = duneggd.Active.NDHPgTPC_v02_Components.NDHPgTPCDetElementBuilder
-HGlayer_builder_name='HGLayerBuilder'
-LGlayer_builder_name='LGLayerBuilder'
-subbuilders	    = [HGlayer_builder_name,LGlayer_builder_name]
-geometry = 'YokeEndcap'
-nsides = 12
-nModules = 5
-output_name = "EndcapYoke"
-pvThickness = Q("20mm")
-yokePhiCutout = Q("90deg")
-yokeThickness = Q("500mm")
-yokeMaterial = "Steel"
-rInnerEcal=Q("2740mm")
-Barrel_halfZ = Q('2600mm')
-nLayers=[4, 16, 4, 31]
-typeLayers=['HG', 'LG', 'HG', 'LG']
